+++ src/calcTimesAll.ts
@@ -42,6 +42,7 @@ export function calcTimesAll(
     Asr: formatTime(raw.Asr),
     Maghrib: formatTime(raw.Maghrib),
     Isha: formatTime(raw.Isha),
+    Midnight: formatTime(raw.Midnight),
     angles: raw.angles,
     Methods,
   };